World: Add stub for PartySystem
